// src/app/models/app-general-quotes-create-dto.ts
export class AppGeneralQuotesCreateDto {


    idCliente: string;
    idDireccionFacturar: number;
    idDireccionEntregar: number;
    idCondPago: number;
    idContacto: number;
    idMtrTipoMoneda: number;
    ordenCompra: string
    fecha: Date;
    observaciones: string;
    usuarioActualiza: string;
    fijarPrecioBs: boolean;
    rif: string;
    razonSocial: string;
    direccion: string;
    idMunicipio: number;

}